requirements.txt
requests~=2.31.0
flask~=3.0.3
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1
Flask-Caching~=2.3.0
redis~=5.0.4